+++ init.rb
@@ -9,7 +9,7 @@ Redmine::Plugin.register :redmine_depending_custom_fields do
   author 'Jan Catrysse'
   description 'Provides depending / cascading custom field formats.'
   url 'https://github.com/jcatrysse/redmine_depending_custom_fields'
-  version '0.0.3'
+  version '0.0.4'
   requires_redmine version_or_higher: '5.0'
 end